pyproject.toml
[project]
name = "stevelandeydotcom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "markdown>=3.8.2",
    "pelican>=4.11.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "markupsafe>=2.1.0",
]
